# src/main/kotlin/com/bank/Multi-currency-banking-api-swagger-01.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:9000","description":"Generated server url"}],"paths":{"/authentication/api/v1/authentication/login":{"post":{"tags":["AuthenticationAPI"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthenticationRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthenticationResponse"}}}}}}},"/api/v1/users/kyc":{"get":{"tags":["KYC-API"],"operationId":"getMyKYC","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"post":{"tags":["KYC-API"],"operationId":"addOrUpdateMyKYC","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/KYCRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/users/accounts":{"get":{"tags":["AccountAPI"],"operationId":"listUserAccounts","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"post":{"tags":["AccountAPI"],"operationId":"createAccount","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAccount"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/users/accounts/{accountNumber}":{"post":{"tags":["AccountAPI"],"operationId":"closeAccount","parameters":[{"name":"accountNumber","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/shop/buy":{"post":{"tags":["ShopAPI"],"operationId":"buyItem","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PurchaseRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/authentication/register":{"post":{"tags":["UserAPI"],"operationId":"registerUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/accounts/withdraw":{"post":{"tags":["TransactionAPI"],"operationId":"withdrawAccount","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WithdrawRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/accounts/transfer":{"post":{"tags":["TransactionAPI"],"operationId":"transferAccounts","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TransferRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/accounts/deposit":{"post":{"tags":["TransactionAPI"],"operationId":"depositAccount","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DepositRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/shop/items/{accountId}":{"get":{"tags":["ShopAPI"],"operationId":"viewItems","parameters":[{"name":"accountId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/shop/history/{accountId}":{"get":{"tags":["ShopAPI"],"operationId":"getShopTransaction","parameters":[{"name":"accountId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/memberships":{"get":{"tags":["MembershipAPI"],"operationId":"getAll","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ListMembershipResponse"}}}}}}}},"/api/v1/memberships/tier/{name}":{"get":{"tags":["MembershipAPI"],"operationId":"getByTierName","parameters":[{"name":"name","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MembershipTierEntity"}}}}}}},"/api/v1/accounts/transactions/{accountId}":{"get":{"tags":["TransactionAPI"],"operationId":"getTransactionHistory","parameters":[{"name":"accountId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}}},"components":{"schemas":{"AuthenticationRequest":{"required":["password","username"],"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}}},"AuthenticationResponse":{"required":["token"],"type":"object","properties":{"token":{"type":"string"}}},"KYCRequest":{"required":["civilId","country","dateOfBirth","firstName","homeAddress","lastName","phoneNumber","salary"],"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"},"dateOfBirth":{"type":"string","format":"date"},"civilId":{"type":"string"},"country":{"type":"string"},"phoneNumber":{"type":"string"},"homeAddress":{"type":"string"},"salary":{"type":"number"}}},"CreateAccount":{"required":["accountType","countryCode","initialBalance"],"type":"object","properties":{"initialBalance":{"type":"number"},"countryCode":{"type":"string"},"accountType":{"type":"string"}}},"PurchaseRequest":{"required":["accountId","itemName"],"type":"object","properties":{"accountId":{"type":"integer","format":"int64"},"itemName":{"type":"string"}}},"CreateUserDTO":{"required":["password","username"],"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}}},"WithdrawRequest":{"required":["accountNumber","amount","countryCode"],"type":"object","properties":{"accountNumber":{"type":"string"},"countryCode":{"type":"string"},"amount":{"type":"number"}}},"TransferRequest":{"required":["amount","countryCode","destinationAccount","sourceAccount"],"type":"object","properties":{"sourceAccount":{"type":"string"},"destinationAccount":{"type":"string"},"amount":{"type":"number"},"countryCode":{"type":"string"}}},"DepositRequest":{"required":["accountNumber","amount","countryCode"],"type":"object","properties":{"accountNumber":{"type":"string"},"countryCode":{"type":"string"},"amount":{"type":"number"}}},"ListMembershipResponse":{"required":["discountAmount","memberLimit","tierName"],"type":"object","properties":{"tierName":{"type":"string"},"memberLimit":{"type":"integer","format":"int32"},"discountAmount":{"type":"number"}}},"MembershipTierEntity":{"required":["discountAmount","memberLimit","tierName"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"tierName":{"type":"string"},"memberLimit":{"type":"integer","format":"int32"},"discountAmount":{"type":"number"}}}}}}